Refactor and improve style for tosca plugin.
