+++ plugins/tosca/src/main/java/de/unistuttgart/iaas/amyassist/amy/plugin/tosca/ToscaLogic.java
@@ -26,7 +26,6 @@ package de.unistuttgart.iaas.amyassist.amy.plugin.tosca;
 import java.util.*;
 
 import org.opentosca.containerapi.client.IOpenTOSCAContainerAPIClient;
-import org.opentosca.containerapi.client.impl.OpenTOSCAContainerLegacyAPIClient;
 import org.opentosca.containerapi.client.model.Application;
 
 import de.unistuttgart.iaas.amyassist.amy.core.configuration.WithDefault;
@@ -34,7 +33,7 @@ import de.unistuttgart.iaas.amyassist.amy.core.di.annotation.PostConstruct;
 import de.unistuttgart.iaas.amyassist.amy.core.di.annotation.Reference;
 import de.unistuttgart.iaas.amyassist.amy.core.di.annotation.Service;
 import de.unistuttgart.iaas.amyassist.amy.core.taskscheduler.api.TaskScheduler;
-import de.unistuttgart.iaas.amyassist.amy.plugin.tosca.configurations.ConfiguartionEntry;
+import de.unistuttgart.iaas.amyassist.amy.plugin.tosca.configurations.ConfigurationEntry;
 import de.unistuttgart.iaas.amyassist.amy.plugin.tosca.configurations.ConfigurationRegistry;
 
 /**
@@ -55,6 +54,9 @@ public class ToscaLogic {
 	@Reference
 	private ConfigurationRegistry registry;
 
+	@Reference
+	private ToscaLibraryAdapter adapter;
+
 	/**
 	 * internal tosca client
 	 */
@@ -64,10 +66,13 @@ public class ToscaLogic {
 	private void connect() {
 		String containerHost = this.configuration.getProperty("CONTAINER_HOST");
 		String containerHostInternal = this.configuration.getProperty("CONTAINER_HOST_INTERNAL");
-		this.apiClient = new OpenTOSCAContainerLegacyAPIClient(containerHost, containerHostInternal);
+		this.apiClient = this.adapter.createLibrary(containerHost, containerHostInternal);
 	}
 
-	public List<ConfiguartionEntry> getConfigurations() {
+	/**
+	 * @return All registered configuration entrys.
+	 */
+	public List<ConfigurationEntry> getConfigurations() {
 		return this.registry.getAll();
 	}
