+++ plugins/tosca/src/main/resources/META-INF/de.unistuttgart.iaas.amyassist.amy.core.di.Services
@@ -1,2 +1,3 @@
+de.unistuttgart.iaas.amyassist.amy.plugin.tosca.ToscaLibaryAdapter
 de.unistuttgart.iaas.amyassist.amy.plugin.tosca.ToscaLogic
 de.unistuttgart.iaas.amyassist.amy.plugin.tosca.configurations.ConfigurationRegistry
